+++ lib/helpers/json_helper.dart
@@ -16,6 +16,22 @@ int jsonGetInt(
   }
 }
 
+double jsonGetDouble(
+  final Json json,
+  final String key, [
+  final double defaultIfNotFound = 0.0,
+]) {
+  final dynamic value = json[key];
+  if (value == null) {
+    return defaultIfNotFound;
+  }
+  try {
+    return value as double;
+  } catch (_) {
+    return defaultIfNotFound;
+  }
+}
+
 bool jsonGetBool(
   final Json json,
   final String key, [
@@ -45,5 +61,21 @@ String jsonGetString(
     return value as String;
   } catch (_) {
     return defaultIfNotFound;
+  }
+}
+
+DateTime jsonGetDate(
+  final Json json,
+  final String key, [
+  final DateTime? defaultIfNotFound,
+]) {
+  final dynamic value = json[key];
+  if (value == null) {
+    return defaultIfNotFound ?? DateTime.parse('19070-01-01');
+  }
+  try {
+    return value as DateTime;
+  } catch (_) {
+    return defaultIfNotFound ?? DateTime.parse('19070-01-01');
   }
 }