+++ lib/models/money_entities/accounts/accounts.dart
@@ -4,6 +4,7 @@ import 'package:money/models/money_entities/accounts/account.dart';
 import 'package:money/models/money_entities/transactions/transaction.dart';
 
 import 'package:money/models/money_entities/money_entity.dart';
+import 'package:uuid/uuid.dart';
 
 class Accounts {
   MoneyObjects<Account> moneyObjects = MoneyObjects<Account>();
@@ -60,37 +61,10 @@ class Accounts {
     return moneyObjects.getAsList();
   }
 
-/*
-0 = "Id"
-1 = "AccountId"
-2 = "OfxAccountId"
-3 = "Name"
-4 = "Description"
-5 = "Type"
-6 = "OpeningBalance"
-7 = "Currency"
-8 = "OnlineAccount"
-9 = "WebSite"
-10 = "ReconcileWarning"
-11 = "LastSync"
-12 = "SyncGuid"
-13 = "Flags"
-14 = "LastBalance"
-15 = "CategoryIdForPrincipal"
-16 = "CategoryIdForInterest"
- */
   load(final List<Map<String, Object?>> rows) async {
     clear();
     for (final Map<String, Object?> row in rows) {
-      final Account a = Account(
-        int.parse(row['Id'].toString()),
-        row['Name'].toString(),
-      );
-      a.accountId = row['AccountId'].toString();
-      a.flags = int.parse(row['Flags'].toString());
-      a.type = AccountType.values[int.parse(row['Type'].toString())];
-      a.openingBalance = double.parse(row['OpeningBalance'].toString());
-
+      final Account a = Account.fromSqlite(row);
       moneyObjects.addEntry(a);
     }
   }
@@ -110,7 +84,25 @@ class Accounts {
       'God-Inc'
     ];
     for (int i = 0; i < names.length; i++) {
-      moneyObjects.addEntry(Account(i, names[i]));
+      moneyObjects.addEntry(Account(
+        id: i,
+        name: names[i],
+        accountId: i.toString(),
+        type: AccountType.checking,
+        description: 'Some description',
+        currency: 'USD',
+        lastSync: DateTime.now(),
+        syncGuid: const Uuid().v4().toString(),
+        flags: 0,
+        openingBalance: 0,
+        ofxAccountId: '',
+        onlineAccount: -1,
+        webSite: '',
+        reconcileWarning: 0,
+        lastBalance: 0.0,
+        categoryIdForPrincipal: 0,
+        categoryIdForInterest: 0,
+      ));
     }
   }