+++ lib/views/view_categories/view_categories.dart
@@ -125,6 +125,8 @@ class ViewCategoriesState extends ViewWidgetState<Category> {
           );
         },
       ),
+      Category.getFieldForDescription(),
+      Category.getFieldForColor(),
       FieldDefinition<Category>(
         name: 'Count',
         type: FieldType.numeric,